non-negligible discrepency between timings within and without hxhim
changed hxhim::Results::duration to be a uint64_t counting nanoseconds

+++ include/hxhim/Results.h
@@ -2,6 +2,7 @@
 #define HXHIM_RESULTS_H
 
 #include <stddef.h>
+#include <stdint.h>
 
 #include "hxhim/constants.h"
 #include "hxhim/struct.h"
@@ -56,7 +57,7 @@ int hxhim_results_valid(hxhim_results_t *res);                             /* wh
 
 // Accessors for the entire list of results
 int hxhim_results_size(hxhim_results_t *res, size_t *size);
-int hxhim_results_duration(hxhim_results_t *res, long double *duration);
+int hxhim_results_duration(hxhim_results_t *res, uint64_t *duration);
 
 // "iterator" functions
 int hxhim_results_valid_iterator(hxhim_results_t *res);                     /* whether or not the pointer is readable */
